tx-generator: Initial check-in of testcase.
A fair amount of the work is on the build system to make it compile
and integrate the code to generate the downrev version's faulty blocks.
A fair amount of the code can likely be trimmed in commits to be folded
into this one to clean up the revision history after this. Dragging in
the entire tx-generator is part of the development process, and the
final result is expected to be just direct calls into block creation
and verification. At the end, only a few modules should be required.
This strategy is insurance to be sure the faulty blocks seen in the
failed attempt at the monadic cleanup are faithfully reproduced.