bench: ci-test bld add modules, pkgs, fix imports
This gets things going to a degree in a ci-test environment. Mostly
packages and modules are added to the cabal file. This commit should
be folded into the initial commit for later merging, never mind the
eventual cutting it down that's supposed to happen. Some minor touchups
were done to refer to modules correctly in the Main stub and use the
right modules instead of Legacy ones for an API change in cardano-api.
The arity of the ProtocolInfoArgsCardano constructor changing still
needs fixing up.